BDOG-516 refactor traits
